removed some warnings for windows-platforms

+++ nselib/pcre.c
@@ -52,7 +52,7 @@ static int get_startoffset(lua_State *L, int stackpos, size_t len)
 	if(startoffset > 0)
 		startoffset--;
 	else if(startoffset < 0) {
-		startoffset += len;
+		startoffset += (int) len;
 		if(startoffset < 0)
 			startoffset = 0;
 	}
